clean structure for auth and rate
